Move ChartType enum to new File
