+++ lib/src/components/chart_type.dart
@@ -1,0 +1,4 @@
+enum ChartType {
+  time,
+  amount,
+}
