+++ lib/time_chart.dart
@@ -2,3 +2,4 @@ library time_chart;
 
 export 'src/time_chart.dart';
 export 'src/components/view_mode.dart';
+export 'src/components/chart_type.dart';
